+++ changelog.md
@@ -1,3 +1,10 @@
+## 1.1.0 (Aug 22, 2013)
+
+ * Add session Restore method
+ * Replaced Init method "login" option with "authenticate"
+ * Init method can now invoke "login", "restore", or "both"
+ * Added improved userId handling
+
 ## 1.0.1 (Aug 22, 2013)
 
  * Improved user data handling